add books operations and admin permission

+++ app/Http/Controllers/AuthorsController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers;
 use App\kotobi\books\Models\Authors;
 use App\kotobi\books\Service\AuthorsService;
 
+
 class AuthorsController extends Controller
 {
 
@@ -17,35 +18,50 @@ class AuthorsController extends Controller
 
     public function index()
     {
+        if(session('group_id') != 1)
+            return redirect('101');
         return view('admin.authors')->with('authors', $this->authorsService->index());
+
     }
 
     public function addForm()
     {
+        if(session('group_id') != 1)
+            return redirect('101');
         return view('admin.addAuthor');
+
     }
 
     public function addProcess()
     {
+        if(session('group_id') != 1)
+            return redirect('101');
         $this->authorsService->add();
         return redirect('authors');
     }
 
     public function delete(Authors $authors)
     {
+        if(session('group_id') != 1)
+            return redirect('101');
+        unlink(public_path('uploads/'.$authors->first()->image));
         $authors->forceDelete();
         return redirect('authors');
     }
 
     public function editForm(Authors $authors)
     {
+        if(session('group_id') != 1)
+            return redirect('101');
         return view('admin.editAuthor')->with('author' ,$authors->findOrFail($authors->author_id));
     }
 
     public function editProcess($id)
     {
-        //dd(request()->get('image'));
+        if(session('group_id') != 1)
+            return redirect('101');
         $this->authorsService->edit($id);
         return redirect('authors');
+
     }
 }
